Split main view into multiple views
