+++ MiniWeather/MiniWeatherApp.swift
@@ -18,8 +18,6 @@ struct MiniWeatherApp: App {
         
         return formatter
     }()
-    /// The location being searched for. This _should_ be in LocationsView but the searchable-related environment values are structured weirdly and this was much easier than changing my view heirarchy a lot.
-    @State private var searchText = ""
     
     init() {
         NSUbiquitousKeyValueStore.default.synchronize()
@@ -27,13 +25,11 @@ struct MiniWeatherApp: App {
 
     var body: some Scene {
         WindowGroup {
-            LocationsView(
-                viewModel: .shared,
-                searchText: $searchText
+            MainView(
+                viewModel: .shared
             )
             .environment(\.timeFormatter, timeFormatter)
             .tint(.primary)
-            .searchable(text: $searchText, prompt: "Search City Name")
         }
         .onChange(of: scenePhase) { oldValue, newValue in
             if oldValue == .background {
