+++ MiniWeather/UI/Location Detail/Views/LocationDetailView.swift
@@ -7,8 +7,12 @@
 
 import SwiftUI
 
-struct LocationDetailView: View {
+struct LocationDetailViewViewModel {
     var location: Location
+}
+
+struct LocationDetailView: View {
+    let viewModel: LocationDetailViewViewModel
     @Binding var weather: WeatherProtocol?
     @Environment(\.colorScheme) var colorScheme
     
@@ -33,9 +37,12 @@ struct LocationDetailView: View {
 
 #Preview {
     LocationDetailView(
-        location: UniversalConstants.location,
+        viewModel:
+            .init(
+                location: UniversalConstants.location
+            ),
         weather: .init(
-            get: { nil },
+            get: { UniversalConstants.weather },
             set: { _ in }
         )
     )
